moved marker components to subfolder
