+++ src/components/Markers/PopupDataRow.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { makeStyles, Theme, createStyles } from '@material-ui/core';
 
 // Import configs
-import STYLES from '../config/styles';
+import STYLES from '../../config/styles';
 
 interface Props
 {
